GRP-931: Rule : problem with ruleThenEnum assignMembershipDisabledDaysForOwnerGroupId wraps integer limit needs long
git-svn-id: https://svn.internet2.edu/svn/i2mi/branches/GROUPER_2_1_BRANCH@8732 bc24b658-cf39-4249-b827-e3526eaaeb42

+++ grouper/src/grouper/edu/internet2/middleware/grouper/app/loader/GrouperLoaderScheduleType.java
@@ -84,7 +84,7 @@ public enum GrouperLoaderScheduleType {
       
       //start time is the interval seconds / 5, rand
       int startSeconds = (int)(Math.random() * intervalSeconds);
-      Date startTime = new Date(System.currentTimeMillis() + (startSeconds*1000));
+      Date startTime = new Date(System.currentTimeMillis() + ((long)startSeconds*1000));
       
       simpleTrigger.setStartTime(startTime);
       return simpleTrigger;
