+++ grouper/src/grouper/edu/internet2/middleware/grouper/externalSubjects/ExternalSubject.java
@@ -1065,7 +1065,7 @@ public class ExternalSubject extends GrouperAPI implements GrouperHasContext,
             group.addMember(subject, false);
             if (expireAfterDays > 0) {
               Membership membership = group.getImmediateMembership(Group.getDefaultList(), subject, true, true);
-              membership.setEnabledTime(new Timestamp(System.currentTimeMillis() + (expireAfterDays*24*60*60*1000)));
+              membership.setEnabledTime(new Timestamp(System.currentTimeMillis() + ((long)expireAfterDays*24*60*60*1000)));
               membership.update();
             }
           }
